Vertex size 3 -> 2, some dummy objects, render Matrix applying, code cleaned

+++ src/org/sevenzip/compression/LZMA/Base.java
@@ -56,7 +56,7 @@ public class Base
 		len -= kMatchMinLen;
 		if (len < kNumLenToPosStates)
 			return len;
-		return (int)(kNumLenToPosStates - 1);
+		return kNumLenToPosStates - 1;
 	}
 	
 	public static final int kNumAlignBits = 4;
